syscall: simplify interface
Macros cast arguments to long automatically.

+++ src/stdio/linux/write.c
@@ -7,5 +7,5 @@
 
 ssize_t write(int fd, const void *buf, size_t count)
 {
-    return syscall(SYS_write, fd, (long)buf, count);
+    return syscall(SYS_write, fd, buf, count);
 }
